# backend/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"LoginResult": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Error": {
				"properties": {
					"message": {
						"type": "string"
					},
					"details": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"message",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponse_LoginResult_": {
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/LoginResult"
							}
						],
						"nullable": true
					},
					"error": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Error"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISignInPayload": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponse_User_": {
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/User"
							}
						],
						"nullable": true
					},
					"error": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Error"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISignUpPayload": {
				"properties": {
					"firstName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"firstName",
					"lastName",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponse_string_": {
				"properties": {
					"data": {
						"type": "string",
						"nullable": true
					},
					"error": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Error"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateCategoryRequest": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryItem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"alias": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"image",
					"alias"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CategoryItems": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/CategoryItem"
						},
						"type": "array"
					}
				},
				"required": [
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponse_CategoryItems_": {
				"properties": {
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/CategoryItems"
							}
						],
						"nullable": true
					},
					"error": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Error"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductItem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"photo": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"image",
					"categoryId",
					"description",
					"photo",
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BaseResponse_ProductItem-Array_": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ProductItem"
						},
						"type": "array",
						"nullable": true
					},
					"error": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Error"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"data",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProductQueryRequest": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"categoryId": {
						"type": "string"
					},
					"limit": {
						"type": "integer",
						"format": "int32",
						"default": "100",
						"minimum": 0
					},
					"offset": {
						"type": "integer",
						"format": "int32",
						"minimum": 0
					}
				},
				"required": [
					"limit",
					"offset"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "fake-market-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/v1/auth/sign-in": {
			"post": {
				"operationId": "SignIn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse_LoginResult_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISignInPayload"
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/sign-up": {
			"post": {
				"operationId": "SignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse_User_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISignUpPayload"
							}
						}
					}
				}
			}
		},
		"/api/v1/category/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse_string_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateCategoryRequest"
							}
						}
					}
				}
			}
		},
		"/api/v1/category/list": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse_CategoryItems_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/api/v1/product/query": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BaseResponse_ProductItem-Array_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProductQueryRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}